fastapi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-community
langchain-google-genai
chromadb
PyPDF2
python-dotenv
python-dateutil